Allow chosing target network
Ss58 serialization is network-dependent
(https://github.com/paritytech/substrate/wiki/External-Address-Format-(SS58))
Therefore, we need to know the target network to output the resulting
address in the expected format.

Also adds a sanity check: all submitted addresses should belong to the
very same network we're asking about. If it is not the case, we just
`panic!`